requirements.txt
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
boto3==1.22.10
botocore==1.25.10
click==8.1.3
DateTime==4.4
decorator==5.1.1
executing==0.8.3
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.11.3
ipython==8.3.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.36
stack-data==0.2.0
traitlets==5.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
zope.interface==5.4.0
